pyproject.toml
[tool.poetry]
name = "llama-bot"
version = "0.1.0"
description = ""
authors = ["Roman Solomatin <36135455+Samoed@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.5"
pydantic-settings = "^2.0.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.20"}
asyncpg = "^0.28.0"
alembic = "^1.12.0"
sqlalchemy-utils = "^0.41.1"
psycopg2-binary = "^2.9.7"
pygments = "^2.16.1"
markupsafe = "^2.1.3"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
pyupgrade = "^3.10.1"
types-requests = "^2.31.0.2"

[tool.isort]
profile = "black"
combine_as_imports = true
skip_gitignore = true
py_version = 310

# https://github.com/tiangolo/sqlmodel/blob/main/pyproject.toml
[tool.mypy]
python_version = "3.10"
# --strict
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
plugins = [
    "pydantic.mypy",
]
ignore_missing_imports = false
# --strict end
exclude = [
    "tests/"
]

[[tool.mypy.overrides]]
module = [
    "sqlalchemy_utils",
    "nltk",
    "pandas",
    "schedule",
    "uvicorn",
]
ignore_missing_imports = true

[tool.black]
line-length = 120
target-version = ['py310']
preview = true
exclude = '''
/(
    .git
  | .venv
  | notebooks
)/
'''